requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Telegram bot
aiogram==3.7.0

# Database
sqlalchemy==2.0.31
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.2

# HTTP client
httpx==0.27.0

# Scheduler
apscheduler==3.10.4
pytz==2024.1

# Environment & config
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4

# Utilities
python-multipart==0.0.9
slowapi==0.1.9

# Logging
structlog==24.2.0
